fix: move s3_get out of subdir
